+++ Web/Edubase.Web.UIUnitTests/Areas/Governors/Controllers/AddOrEdit/AddGovernorControllerTests.cs
@@ -1,154 +1,150 @@
 using System;
 using System.Collections.Generic;
-using System.Collections.Specialized;
 using System.Linq;
 using System.Security.Principal;
-using System.Text;
 using System.Threading.Tasks;
 using System.Web;
 using System.Web.Mvc;
 using System.Web.Routing;
 using Edubase.Services.Domain;
 using Edubase.Services.Enums;
-using Edubase.Services.Establishments;
 using Edubase.Services.Establishments.Models;
 using Edubase.Services.Exceptions;
-using Edubase.Services.Governors;
 using Edubase.Services.Governors.DisplayPolicies;
 using Edubase.Services.Governors.Models;
-using Edubase.Services.Groups;
 using Edubase.Services.Groups.Models;
-using Edubase.Services.Lookup;
-using Edubase.Web.UI.Areas;
-using Edubase.Web.UI.Areas.Governors.Controllers;
 using Edubase.Web.UI.Areas.Governors.Models;
-using Edubase.Web.UI.Helpers;
 using Edubase.Web.UI.Models;
+using Edubase.Web.UIUnitTests.Areas.Governors.TestBase.Edubase.Web.UIUnitTests.Areas.Governors.TestBase;
 using Moq;
 using Xunit;
 
 namespace Edubase.Web.UIUnitTests.Areas.Governors.Controllers.AddOrEdit
 {
-    public class AddGovernorControllerTests : IDisposable
+    public class AddGovernorControllerTests : GovernorControllerTestBase
     {
-        private readonly GovernorController controller;
-        private readonly Mock<ICachedLookupService> mockCachedLookupService;
-
-        private readonly Mock<IGovernorsReadService> mockGovernorsReadService = new Mock<IGovernorsReadService>(MockBehavior.Strict);
-        private readonly Mock<IGovernorsWriteService> mockGovernorsWriteService = new Mock<IGovernorsWriteService>(MockBehavior.Strict);
-        private readonly Mock<IGroupReadService> mockGroupReadService = new Mock<IGroupReadService>(MockBehavior.Strict);
-        private readonly Mock<IEstablishmentReadService> mockEstablishmentReadService = new Mock<IEstablishmentReadService>(MockBehavior.Strict);
-        private readonly Mock<ILayoutHelper> mockLayoutHelper = new Mock<ILayoutHelper>(MockBehavior.Strict);
-        private readonly Mock<UrlHelper> mockUrlHelper = new Mock<UrlHelper>(MockBehavior.Loose);
-        private readonly Mock<ControllerContext> mockControllerContext = new Mock<ControllerContext>(MockBehavior.Strict);
-        private readonly Mock<HttpRequestBase> mockHttpRequestBase = new Mock<HttpRequestBase>(MockBehavior.Strict);
-        private readonly Mock<HttpContextBase> mockHttpContextBase = new Mock<HttpContextBase>(MockBehavior.Strict);
-        private readonly Mock<IPrincipal> mockPrincipal = new Mock<IPrincipal>(MockBehavior.Strict);
-        private readonly Mock<IIdentity> mockIdentity = new Mock<IIdentity>(MockBehavior.Strict);
-        private readonly Mock<IGovernorsGridViewModelFactory> mockGovernorGridViewModelFactory = new Mock<IGovernorsGridViewModelFactory>(MockBehavior.Strict);
-
-        private bool disposedValue;
-
-        public AddGovernorControllerTests()
-        {
-            mockCachedLookupService = MockHelper.SetupCachedLookupService();
-
-            mockEstablishmentReadService.Setup(e => e.GetEstabType2EducationPhaseMap())
-                .Returns(new Dictionary<eLookupEstablishmentType, eLookupEducationPhase[]>());
-
-            mockUrlHelper.Setup(u => u.RouteUrl(It.IsAny<string>(), It.IsAny<object>()))
-                .Returns("fake url");
-
-            controller = new GovernorController(
-                mockGovernorsReadService.Object,
-                mockCachedLookupService.Object,
-                mockGovernorsWriteService.Object,
-                mockGroupReadService.Object,
-                mockEstablishmentReadService.Object,
-                mockLayoutHelper.Object);
-
-            SetupController();
-        }
-
-        protected void SetupController()
-        {
-            SetupHttpRequest();
-            controller.ControllerContext = mockControllerContext.Object;
-            mockControllerContext.SetupGet(c => c.Controller).Returns(controller);
-            controller.Url = mockUrlHelper.Object;
-        }
-
-        private void SetupHttpRequest()
-        {
-            mockHttpRequestBase.SetupGet(x => x.QueryString)
-                .Returns(HttpUtility.ParseQueryString(string.Empty));
-            mockHttpContextBase.SetupGet(x => x.Request)
-                .Returns(mockHttpRequestBase.Object);
-            mockHttpContextBase.SetupGet(x => x.User)
-                .Returns(mockPrincipal.Object);
-            mockControllerContext.SetupGet(x => x.HttpContext)
-                .Returns(mockHttpContextBase.Object);
-            mockControllerContext.SetupGet(x => x.IsChildAction)
-                .Returns(false);
-            mockControllerContext.SetupGet(x => x.RouteData)
-                .Returns(new System.Web.Routing.RouteData());
-            mockPrincipal.SetupGet(x => x.Identity)
-                .Returns(mockIdentity.Object);
-        }
-
-        [Fact()]
+        [Fact]
         public async Task Gov_AddEditOrReplace_NullParams()
         {
-            mockControllerContext.SetupGet(c => c.RouteData).Returns(new RouteData(new Route("", new PageRouteHandler("~/")), new PageRouteHandler("~/")));
-            await Assert.ThrowsAsync<EdubaseException>(() => controller.AddEditOrReplace(null, null, null, null));
+            // Arrange
+            var controller = BuildController();
+
+            var route = new Route("", new PageRouteHandler("~/"));
+            var rd = new RouteData(route, new PageRouteHandler("~/"));
+            rd.Values["controller"] = "Governors";
+            rd.Values["action"] = "AddEditOrReplace";
+
+            controller.ControllerContext = new ControllerContext(
+                controller.ControllerContext.HttpContext,
+                rd,
+                controller);
+
+            // Act & Assert
+            await Assert.ThrowsAsync<EdubaseException>(() =>
+                controller.AddEditOrReplace(null, null, null, null));
         }
 
-        [Fact()]
+        [Fact]
         public async Task Gov_AddEditOrReplace_RoleSpecified_Single_NotShared_AlreadyExists()
         {
+            // Arrange
             var estabUrn = 4;
-            mockGovernorsReadService.Setup(g => g.GetGovernorListAsync(estabUrn, null, It.IsAny<IPrincipal>())).ReturnsAsync(() => new GovernorsDetailsDto
-            {
-                CurrentGovernors = new List<GovernorModel> { new GovernorModel { RoleId = (int) eLookupGovernorRole.ChairOfGovernors } }
-            });
-            mockControllerContext.SetupGet(c => c.RouteData).Returns(new RouteData(new Route("", new PageRouteHandler("~/")), new PageRouteHandler("~/")));
+            var controller = BuildController();
 
-            var result = await controller.AddEditOrReplace(null, estabUrn, eLookupGovernorRole.ChairOfGovernors, null);
+            mockGovernorsReadService
+                .Setup(g => g.GetGovernorListAsync(estabUrn, null, It.IsAny<IPrincipal>()))
+                .ReturnsAsync(() => new GovernorsDetailsDto
+                {
+                    CurrentGovernors = new List<GovernorModel>
+                    {
+                    new GovernorModel { RoleId = (int)eLookupGovernorRole.ChairOfGovernors }
+                    }
+                });
 
-            var redirectResult = result as RedirectToRouteResult;
-            Assert.NotNull(redirectResult);
+            var route = new Route("", new PageRouteHandler("~/"));
+            var routeData = new RouteData(route, new PageRouteHandler("~/"));
+            routeData.Values["controller"] = "Governors";
+            routeData.Values["action"] = "AddEditOrReplace";
+
+            controller.ControllerContext = new ControllerContext(
+                controller.ControllerContext.HttpContext,
+                routeData,
+                controller);
+
+            // Act
+            var result = await controller.AddEditOrReplace(
+                groupUId: null,
+                establishmentUrn: estabUrn,
+                role: eLookupGovernorRole.ChairOfGovernors,
+                gid: null);
+
+            // Assert
+            var redirectResult = Assert.IsType<RedirectToRouteResult>(result);
             Assert.Equal("EstabEditGovernance", redirectResult.RouteName);
         }
 
-        [Fact()]
+        [Fact]
         public async Task Gov_AddEditOrReplace_RoleSpecified_Single_NotShared_DoesntExist()
         {
+            // Arrange
             var estabUrn = 4;
-            mockGovernorsReadService.Setup(g => g.GetGovernorListAsync(estabUrn, null, It.IsAny<IPrincipal>())).ReturnsAsync(() => new GovernorsDetailsDto
-            {
-                CurrentGovernors = new List<GovernorModel>()
-            });
-            mockControllerContext.SetupGet(c => c.RouteData).Returns(new RouteData(new Route("", new PageRouteHandler("~/")), new PageRouteHandler("~/")));
-            mockLayoutHelper.Setup(l => l.PopulateLayoutProperties(It.IsAny<CreateEditGovernorViewModel>(), estabUrn, null, It.IsAny<IPrincipal>(), It.IsAny<Action<EstablishmentModel>>(), It.IsAny<Action<GroupModel>>())).Returns(Task.CompletedTask);
-            mockGovernorsReadService.Setup(g => g.GetEditorDisplayPolicyAsync(eLookupGovernorRole.ChairOfGovernors, false, It.IsAny<IPrincipal>())).ReturnsAsync(() => new GovernorDisplayPolicy());
+            var controller = BuildController();
 
-            var result = await controller.AddEditOrReplace(null, estabUrn, eLookupGovernorRole.ChairOfGovernors, null);
+            var route = new Route("", new PageRouteHandler("~/"));
+            var rd = new RouteData(route, new PageRouteHandler("~/"));
+            controller.ControllerContext = new ControllerContext(
+                controller.ControllerContext.HttpContext,
+                rd,
+                controller);
 
-            var viewResult = result as ViewResult;
-            Assert.NotNull(viewResult);
+            SetupCommonLookupMocks();
 
-            var model = viewResult.Model as CreateEditGovernorViewModel;
-            Assert.NotNull(model);
+            mockGovernorsReadService
+                .Setup(g => g.GetGovernorListAsync(estabUrn, null, It.IsAny<IPrincipal>()))
+                .ReturnsAsync(() => new GovernorsDetailsDto
+                {
+                    CurrentGovernors = new List<GovernorModel>()
+                });
+
+            mockLayoutHelper
+                .Setup(l => l.PopulateLayoutProperties(
+                    It.IsAny<CreateEditGovernorViewModel>(),
+                    estabUrn,
+                    null,
+                    It.IsAny<IPrincipal>(),
+                    It.IsAny<Action<EstablishmentModel>>(),
+                    It.IsAny<Action<GroupModel>>()))
+                .Returns(Task.CompletedTask);
+
+            mockGovernorsReadService
+                .Setup(g => g.GetEditorDisplayPolicyAsync(
+                    eLookupGovernorRole.ChairOfGovernors,
+                    false,
+                    It.IsAny<IPrincipal>()))
+                .ReturnsAsync(() => new GovernorDisplayPolicy());
+
+            // Act
+            var result = await controller.AddEditOrReplace(
+                groupUId: null,
+                establishmentUrn: estabUrn,
+                role: eLookupGovernorRole.ChairOfGovernors,
+                gid: null);
+
+            // Assert
+            var viewResult = Assert.IsType<ViewResult>(result);
+            var model = Assert.IsType<CreateEditGovernorViewModel>(viewResult.Model);
             Assert.Equal(eLookupGovernorRole.ChairOfGovernors, model.GovernorRole);
             Assert.Equal(estabUrn, model.EstablishmentUrn);
             Assert.Null(model.GroupUId);
         }
 
-        [Fact()]
+        [Fact]
         public async Task Gov_AddEditOrReplace_GIDSpecified()
         {
+            // Arrange
             var estabUrn = 4;
             var governorId = 1032;
+            var controller = BuildController();
 
             var governor = new GovernorModel
             {
@@ -156,35 +152,84 @@ namespace Edubase.Web.UIUnitTests.Areas.Governors.Controllers.AddOrEdit
                 RoleId = (int) eLookupGovernorRole.Governor
             };
 
-            mockControllerContext.SetupGet(c => c.RouteData).Returns(new RouteData(new Route("", new PageRouteHandler("~/")), new PageRouteHandler("~/")));
-            mockGovernorsReadService.Setup(g => g.GetGovernorAsync(governorId, It.IsAny<IPrincipal>())).ReturnsAsync(() => governor);
-            mockLayoutHelper.Setup(l => l.PopulateLayoutProperties(It.IsAny<CreateEditGovernorViewModel>(), estabUrn, null, It.IsAny<IPrincipal>(), It.IsAny<Action<EstablishmentModel>>(), It.IsAny<Action<GroupModel>>())).Returns(Task.CompletedTask);
-            mockGovernorsReadService.Setup(g => g.GetEditorDisplayPolicyAsync(eLookupGovernorRole.Governor, false, It.IsAny<IPrincipal>())).ReturnsAsync(() => new GovernorDisplayPolicy());
+            var route = new Route("", new PageRouteHandler("~/"));
+            var rd = new RouteData(route, new PageRouteHandler("~/"));
+            controller.ControllerContext = new ControllerContext(
+                controller.ControllerContext.HttpContext,
+                rd,
+                controller);
 
-            var result = await controller.AddEditOrReplace(null, estabUrn, null, 1032);
+            SetupCommonLookupMocks();
 
-            var viewResult = result as ViewResult;
-            Assert.NotNull(viewResult);
+            mockGovernorsReadService
+                .Setup(g => g.GetGovernorAsync(governorId, It.IsAny<IPrincipal>()))
+                .ReturnsAsync(() => governor);
 
-            var model = viewResult.Model as CreateEditGovernorViewModel;
-            Assert.NotNull(model);
+            mockLayoutHelper
+                .Setup(l => l.PopulateLayoutProperties(
+                    It.IsAny<CreateEditGovernorViewModel>(),
+                    estabUrn,
+                    null,
+                    It.IsAny<IPrincipal>(),
+                    It.IsAny<Action<EstablishmentModel>>(),
+                    It.IsAny<Action<GroupModel>>()))
+                .Returns(Task.CompletedTask);
+
+            mockGovernorsReadService
+                .Setup(g => g.GetEditorDisplayPolicyAsync(
+                    eLookupGovernorRole.Governor,
+                    false,
+                    It.IsAny<IPrincipal>()))
+                .ReturnsAsync(() => new GovernorDisplayPolicy());
+
+            // Act
+            var result = await controller.AddEditOrReplace(
+                groupUId: null,
+                establishmentUrn: estabUrn,
+                role: null,
+                gid: governorId);
+
+            // Assert
+            var viewResult = Assert.IsType<ViewResult>(result);
+            var model = Assert.IsType<CreateEditGovernorViewModel>(viewResult.Model);
             Assert.Equal(CreateEditGovernorViewModel.EditMode.Edit, model.Mode);
         }
 
-        [Fact()]
+        [Fact]
         public async Task Gov_AddEditOrReplace_RoleSpecified_Shared()
         {
+            // Arrange
             var estabUrn = 4;
-            mockGovernorsReadService.Setup(g => g.GetGovernorListAsync(estabUrn, null, It.IsAny<IPrincipal>())).ReturnsAsync(() => new GovernorsDetailsDto
-            {
-                CurrentGovernors = new List<GovernorModel>()
-            });
-            mockControllerContext.SetupGet(c => c.RouteData).Returns(new RouteData(new Route("", new PageRouteHandler("~/")), new PageRouteHandler("~/")));
+            var controller = BuildController();
 
-            var result = await controller.AddEditOrReplace(null, estabUrn, eLookupGovernorRole.Establishment_SharedChairOfLocalGoverningBody, null);
+            var route = new Route("", new PageRouteHandler("~/"));
+            var routeData = new RouteData(route, new PageRouteHandler("~/"));
+            routeData.Values["controller"] = "Governors";
+            routeData.Values["action"] = "AddEditOrReplace";
 
-            var redirectResult = result as RedirectToRouteResult;
-            Assert.NotNull(redirectResult);
+            controller.ControllerContext = new ControllerContext(
+                    controller.ControllerContext.HttpContext,
+                    routeData,
+                    controller);
+
+            SetupCommonLookupMocks();
+
+            mockGovernorsReadService
+                .Setup(g => g.GetGovernorListAsync(estabUrn, null, It.IsAny<IPrincipal>()))
+                .ReturnsAsync(() => new GovernorsDetailsDto
+                {
+                    CurrentGovernors = new List<GovernorModel>()
+                });
+
+            // Act
+            var result = await controller.AddEditOrReplace(
+                groupUId: null,
+                establishmentUrn: estabUrn,
+                role: eLookupGovernorRole.Establishment_SharedChairOfLocalGoverningBody,
+                gid: null);
+
+            // Assert
+            var redirectResult = Assert.IsType<RedirectToRouteResult>(result);
             Assert.Equal("SelectSharedGovernor", redirectResult.RouteName);
         }
 
@@ -195,21 +240,24 @@ namespace Edubase.Web.UIUnitTests.Areas.Governors.Controllers.AddOrEdit
             var estabUrn = 123;
             var gid = 456;
 
-            // Make the route URL contain "/Replace/" so replaceMode == true
-            mockControllerContext
-                .SetupGet(c => c.RouteData)
-                .Returns(new RouteData(
-                    new Route("establishment/edit/{establishmentUrn}/governance/Replace/{gid}",
-                              new PageRouteHandler("~/")),
-                    new PageRouteHandler("~/")));
-
             var qs = HttpUtility.ParseQueryString(string.Empty);
             qs["d"] = "10";
             qs["m"] = "12";
             qs["y"] = "2024";
-            mockHttpRequestBase
-                .SetupGet(r => r.QueryString)
-                .Returns(qs);
+
+            var controller = BuildController(qs);
+
+            var route = new Route(
+                "establishment/edit/{establishmentUrn}/governance/Replace/{gid}",
+                new PageRouteHandler("~/"));
+            var routeData = new RouteData(route, new PageRouteHandler("~/"));
+            routeData.Values["establishmentUrn"] = estabUrn;
+            routeData.Values["gid"] = gid;
+
+            controller.ControllerContext = new ControllerContext(
+                controller.ControllerContext.HttpContext,
+                routeData,
+                controller);
 
             var previousGov = new GovernorModel
             {
@@ -231,7 +279,6 @@ namespace Edubase.Web.UIUnitTests.Areas.Governors.Controllers.AddOrEdit
             // Assert
             var redirect = Assert.IsType<RedirectToRouteResult>(result);
             Assert.Equal("EstabReplaceChair", redirect.RouteName);
-
             Assert.Equal(estabUrn, redirect.RouteValues["establishmentUrn"]);
             Assert.Equal(gid, redirect.RouteValues["gid"]);
             Assert.Equal("10", redirect.RouteValues["d"]);
@@ -245,6 +292,7 @@ namespace Edubase.Web.UIUnitTests.Areas.Governors.Controllers.AddOrEdit
         {
             // Arrange
             var groupId = 42;
+            var controller = BuildController();
 
             var dto = new GovernorsDetailsDto
             {
@@ -260,11 +308,17 @@ namespace Edubase.Web.UIUnitTests.Areas.Governors.Controllers.AddOrEdit
                 .Setup(g => g.GetGovernorListAsync(null, groupId, It.IsAny<IPrincipal>()))
                 .ReturnsAsync(dto);
 
-            mockControllerContext
-                .SetupGet(c => c.RouteData)
-                .Returns(new RouteData(
-                    new Route("group/edit/{groupUId}/governance", new PageRouteHandler("~/")),
-                    new PageRouteHandler("~/")));
+            var route = new Route("", new PageRouteHandler("~/"));
+            var routeData = new RouteData(route, new PageRouteHandler("~/"));
+            routeData.Values["controller"] = "Governors";
+            routeData.Values["action"] = "AddEditOrReplace";
+
+            controller.ControllerContext = new ControllerContext(
+                controller.ControllerContext.HttpContext,
+                routeData,
+                controller);
+
+            SetupCommonLookupMocks();
 
             mockLayoutHelper
                 .Setup(l => l.PopulateLayoutProperties(
@@ -286,7 +340,6 @@ namespace Edubase.Web.UIUnitTests.Areas.Governors.Controllers.AddOrEdit
             // Assert
             var redirect = Assert.IsType<RedirectToRouteResult>(result);
             Assert.Equal("GroupEditGovernance", redirect.RouteName);
-
             Assert.Equal(groupId, redirect.RouteValues["groupUId"]);
             Assert.Null(redirect.RouteValues["establishmentUrn"]);
             Assert.True((bool) redirect.RouteValues["roleAlreadyExists"]);
@@ -301,22 +354,23 @@ namespace Edubase.Web.UIUnitTests.Areas.Governors.Controllers.AddOrEdit
             var gid = 200;
             var replacementGovernorId = 300;
 
-            // Force replaceMode = true by setting a route URL with "/Replace/" in it
-            mockControllerContext
-                .SetupGet(c => c.RouteData)
-                .Returns(new RouteData(
-                    new Route("establishment/edit/{establishmentUrn}/governance/Replace/{gid}",
-                              new PageRouteHandler("~/")),
-                    new PageRouteHandler("~/")));
-
-            // Provide query string with gid2 = replacementGovernorId
             var qs = HttpUtility.ParseQueryString(string.Empty);
             qs["gid2"] = replacementGovernorId.ToString();
-            mockHttpRequestBase
-                .SetupGet(r => r.QueryString)
-                .Returns(qs);
 
-            // The governor being edited/replaced (a Chair of Trustees)
+            var controller = BuildController(qs);
+
+            var route = new Route(
+                "establishment/edit/{establishmentUrn}/governance/Replace/{gid}",
+                new PageRouteHandler("~/"));
+            var routeData = new RouteData(route, new PageRouteHandler("~/"));
+            routeData.Values["establishmentUrn"] = estabUrn;
+            routeData.Values["gid"] = gid;
+
+            controller.ControllerContext = new ControllerContext(
+                controller.ControllerContext.HttpContext,
+                routeData,
+                controller);
+
             var model = new GovernorModel
             {
                 Id = gid,
@@ -331,41 +385,37 @@ namespace Edubase.Web.UIUnitTests.Areas.Governors.Controllers.AddOrEdit
                 .Setup(g => g.GetGovernorAsync(gid, It.IsAny<IPrincipal>()))
                 .ReturnsAsync(model);
 
-            // The list of existing governors/trustees available to be selected as replacement
             var dto = new GovernorsDetailsDto
             {
                 CurrentGovernors = new List<GovernorModel>
-        {
-            new GovernorModel
             {
-                Id = replacementGovernorId,
-                RoleId = (int)eLookupGovernorRole.Governor,
-                Person_FirstName = "Beth",
-                Person_LastName = "Gov",
-                AppointmentStartDate = new DateTime(2020, 1, 1),
-                AppointmentEndDate   = new DateTime(2024, 1, 1),
-                DOB = new DateTime(1980, 1, 1),
-                Person_TitleId = 1
+                new GovernorModel
+                {
+                    Id = replacementGovernorId,
+                    RoleId = (int)eLookupGovernorRole.Governor,
+                    Person_FirstName = "Beth",
+                    Person_LastName = "Gov",
+                    AppointmentStartDate = new DateTime(2020, 1, 1),
+                    AppointmentEndDate = new DateTime(2024, 1, 1),
+                    DOB = new DateTime(1980, 1, 1),
+                    Person_TitleId = 1
+                },
+                new GovernorModel
+                {
+                    Id = 301,
+                    RoleId = (int)eLookupGovernorRole.Trustee,
+                    Person_FirstName = "Chris",
+                    Person_LastName = "Trustee",
+                    Person_TitleId = 1
+                }
             },
-            new GovernorModel
-            {
-                Id = 301,
-                RoleId = (int)eLookupGovernorRole.Trustee,
-                Person_FirstName = "Chris",
-                Person_LastName = "Trustee",
-                Person_TitleId = 1
-            }
-        },
                 HistoricalGovernors = new List<GovernorModel>()
             };
 
-            // Used twice: once inside RoleAllowed, once inside the replace block
             mockGovernorsReadService
                 .Setup(g => g.GetGovernorListAsync(estabUrn, null, It.IsAny<IPrincipal>()))
                 .ReturnsAsync(dto);
 
-            // Allow the ChairOfTrustees role (RoleAllowed should return true)
-            // by not having conflicting roles in dto.CurrentGovernors (we only have Governor/Trustee)
             mockGovernorsReadService
                 .Setup(g => g.GetEditorDisplayPolicyAsync(
                     eLookupGovernorRole.ChairOfTrustees,
@@ -383,7 +433,9 @@ namespace Edubase.Web.UIUnitTests.Areas.Governors.Controllers.AddOrEdit
                     It.IsAny<Action<GroupModel>>()))
                 .Returns(Task.CompletedTask);
 
-            // Act: establishment context, replace mode (via route), ChairOfTrustees role, with gid
+            SetupCommonLookupMocks();
+
+            // Act
             var result = await controller.AddEditOrReplace(
                 groupUId: null,
                 establishmentUrn: estabUrn,
@@ -394,27 +446,20 @@ namespace Edubase.Web.UIUnitTests.Areas.Governors.Controllers.AddOrEdit
             var viewResult = Assert.IsType<ViewResult>(result);
             var vm = Assert.IsType<CreateEditGovernorViewModel>(viewResult.Model);
 
-            // We are in Replace mode, not Create/Edit
             Assert.Equal(CreateEditGovernorViewModel.EditMode.Replace, vm.Mode);
-
-            // ReplaceGovernorViewModel should be populated from the chair model
             Assert.Equal(gid, vm.ReplaceGovernorViewModel.GID);
             Assert.NotNull(vm.ReplaceGovernorViewModel.AppointmentEndDate);
             Assert.Equal(model.AppointmentEndDate, vm.ReplaceGovernorViewModel.AppointmentEndDate.ToDateTime());
             Assert.False(string.IsNullOrWhiteSpace(vm.ReplaceGovernorViewModel.Name));
 
-            // Existing governors list built from governorsOrTrustees (Governor + Trustee)
             Assert.NotNull(vm.ExistingGovernors);
             var existingList = vm.ExistingGovernors.ToList();
             Assert.Equal(2, existingList.Count);
 
-            // The replacement governor from gid2 should be selected and prepopulated
             var selectedItem = Assert.Single(existingList.Where(i => i.Selected));
             Assert.Equal(replacementGovernorId.ToString(), selectedItem.Value);
-
             Assert.NotNull(vm.SelectedGovernor);
             Assert.Equal(replacementGovernorId, vm.SelectedGovernor.Id);
-            // PrepopulateFields should have copied basic fields from SelectedGovernor
             Assert.Equal(vm.SelectedGovernor.Person_FirstName, vm.FirstName);
             Assert.Equal(vm.SelectedGovernor.Person_LastName, vm.LastName);
             Assert.Equal(vm.SelectedGovernor.Person_TitleId, vm.GovernorTitleId);
@@ -428,15 +473,22 @@ namespace Edubase.Web.UIUnitTests.Areas.Governors.Controllers.AddOrEdit
             var gid = 200;
             var replacementId = 300;
 
-            mockControllerContext.SetupGet(c => c.RouteData)
-                .Returns(new RouteData(
-                    new Route("establishment/{establishmentUrn}/governance/Replace/{gid}",
-                              new PageRouteHandler("~/")),
-                    new PageRouteHandler("~/")));
-
             var qs = HttpUtility.ParseQueryString(string.Empty);
             qs["gid2"] = replacementId.ToString();
-            mockHttpRequestBase.SetupGet(r => r.QueryString).Returns(qs);
+
+            var controller = BuildController(qs);
+
+            var route = new Route(
+                "establishment/{establishmentUrn}/governance/Replace/{gid}",
+                new PageRouteHandler("~/"));
+            var routeData = new RouteData(route, new PageRouteHandler("~/"));
+            routeData.Values["establishmentUrn"] = estabUrn;
+            routeData.Values["gid"] = gid;
+
+            controller.ControllerContext = new ControllerContext(
+                controller.ControllerContext.HttpContext,
+                routeData,
+                controller);
 
             var chair = new GovernorModel
             {
@@ -455,25 +507,25 @@ namespace Edubase.Web.UIUnitTests.Areas.Governors.Controllers.AddOrEdit
             var dto = new GovernorsDetailsDto
             {
                 CurrentGovernors = new List<GovernorModel>
-        {
-            new GovernorModel
             {
-                Id = replacementId,
-                RoleId = (int)eLookupGovernorRole.Governor,
-                Person_FirstName = "Beth",
-                Person_LastName = "Gov",
-                Person_TitleId = 1,
-                DOB = new DateTime(1980,1,1)
+                new GovernorModel
+                {
+                    Id = replacementId,
+                    RoleId = (int)eLookupGovernorRole.Governor,
+                    Person_FirstName = "Beth",
+                    Person_LastName = "Gov",
+                    Person_TitleId = 1,
+                    DOB = new DateTime(1980, 1, 1)
+                },
+                new GovernorModel
+                {
+                    Id = 301,
+                    RoleId = (int)eLookupGovernorRole.Trustee,
+                    Person_FirstName = "Chris",
+                    Person_LastName = "Trustee",
+                    Person_TitleId = 1
+                }
             },
-            new GovernorModel
-            {
-                Id = 301,
-                RoleId = (int)eLookupGovernorRole.Trustee,
-                Person_FirstName = "Chris",
-                Person_LastName = "Trustee",
-                Person_TitleId = 1
-            }
-        },
                 HistoricalGovernors = new List<GovernorModel>()
             };
 
@@ -498,6 +550,8 @@ namespace Edubase.Web.UIUnitTests.Areas.Governors.Controllers.AddOrEdit
                     It.IsAny<Action<GroupModel>>()))
                 .Returns(Task.CompletedTask);
 
+            SetupCommonLookupMocks();
+
             // Act
             var result = await controller.AddEditOrReplace(
                 groupUId: null,
@@ -508,7 +562,6 @@ namespace Edubase.Web.UIUnitTests.Areas.Governors.Controllers.AddOrEdit
             // Assert
             var view = Assert.IsType<ViewResult>(result);
             var vm = Assert.IsType<CreateEditGovernorViewModel>(view.Model);
-
             Assert.Equal(CreateEditGovernorViewModel.EditMode.Replace, vm.Mode);
 
             var existing = vm.ExistingGovernors.ToList();
@@ -516,7 +569,6 @@ namespace Edubase.Web.UIUnitTests.Areas.Governors.Controllers.AddOrEdit
 
             var selected = Assert.Single(existing.Where(x => x.Selected));
             Assert.Equal(replacementId.ToString(), selected.Value);
-
             Assert.NotNull(vm.SelectedGovernor);
             Assert.Equal("Beth", vm.FirstName);
             Assert.Equal("Gov", vm.LastName);
@@ -530,14 +582,20 @@ namespace Edubase.Web.UIUnitTests.Areas.Governors.Controllers.AddOrEdit
             var estabUrn = 4002;
             var gid = 201;
 
-            mockControllerContext.SetupGet(c => c.RouteData)
-                .Returns(new RouteData(
-                    new Route("establishment/{establishmentUrn}/governance/Replace/{gid}",
-                              new PageRouteHandler("~/")),
-                    new PageRouteHandler("~/")));
+            var qs = HttpUtility.ParseQueryString(string.Empty);
+            var controller = BuildController(qs);
 
-            mockHttpRequestBase.SetupGet(r => r.QueryString)
-                .Returns(HttpUtility.ParseQueryString(string.Empty));
+            var route = new Route(
+                "establishment/{establishmentUrn}/governance/Replace/{gid}",
+                new PageRouteHandler("~/"));
+            var routeData = new RouteData(route, new PageRouteHandler("~/"));
+            routeData.Values["establishmentUrn"] = estabUrn;
+            routeData.Values["gid"] = gid;
+
+            controller.ControllerContext = new ControllerContext(
+                controller.ControllerContext.HttpContext,
+                routeData,
+                controller);
 
             var chair = new GovernorModel
             {
@@ -548,6 +606,7 @@ namespace Edubase.Web.UIUnitTests.Areas.Governors.Controllers.AddOrEdit
                 AppointmentEndDate = new DateTime(2026, 1, 31),
                 Person_TitleId = 1
             };
+
             mockGovernorsReadService
                 .Setup(g => g.GetGovernorAsync(gid, It.IsAny<IPrincipal>()))
                 .ReturnsAsync(chair);
@@ -555,24 +614,24 @@ namespace Edubase.Web.UIUnitTests.Areas.Governors.Controllers.AddOrEdit
             var dto = new GovernorsDetailsDto
             {
                 CurrentGovernors = new List<GovernorModel>
-        {
-            new GovernorModel
             {
-                Id = 1001,
-                RoleId = (int)eLookupGovernorRole.Governor,
-                Person_FirstName = "Donna",
-                Person_LastName = "Gov",
-                Person_TitleId = 1
+                new GovernorModel
+                {
+                    Id = 1001,
+                    RoleId = (int)eLookupGovernorRole.Governor,
+                    Person_FirstName = "Donna",
+                    Person_LastName = "Gov",
+                    Person_TitleId = 1
+                },
+                new GovernorModel
+                {
+                    Id = 1002,
+                    RoleId = (int)eLookupGovernorRole.Trustee,
+                    Person_FirstName = "Elliot",
+                    Person_LastName = "Trustee",
+                    Person_TitleId = 1
+                }
             },
-            new GovernorModel
-            {
-                Id = 1002,
-                RoleId = (int)eLookupGovernorRole.Trustee,
-                Person_FirstName = "Elliot",
-                Person_LastName = "Trustee",
-                Person_TitleId = 1
-            }
-        },
                 HistoricalGovernors = new List<GovernorModel>()
             };
 
@@ -597,6 +656,8 @@ namespace Edubase.Web.UIUnitTests.Areas.Governors.Controllers.AddOrEdit
                     It.IsAny<Action<GroupModel>>()))
                 .Returns(Task.CompletedTask);
 
+            SetupCommonLookupMocks();
+
             // Act
             var result = await controller.AddEditOrReplace(
                 groupUId: null,
@@ -607,12 +668,10 @@ namespace Edubase.Web.UIUnitTests.Areas.Governors.Controllers.AddOrEdit
             // Assert
             var view = Assert.IsType<ViewResult>(result);
             var vm = Assert.IsType<CreateEditGovernorViewModel>(view.Model);
-
             Assert.Equal(CreateEditGovernorViewModel.EditMode.Replace, vm.Mode);
 
             var existing = vm.ExistingGovernors.ToList();
             Assert.Equal(2, existing.Count);
-
             Assert.DoesNotContain(existing, x => x.Selected);
         }
 
@@ -622,17 +681,20 @@ namespace Edubase.Web.UIUnitTests.Areas.Governors.Controllers.AddOrEdit
             // Arrange
             var estabUrn = 7777;
             var gid = 123;
+            var controller = BuildController();
 
-            mockControllerContext.SetupGet(c => c.RouteData)
-                .Returns(new RouteData(
-                    new Route("establishment/{urn}/governance", new PageRouteHandler("~/")),
-                    new PageRouteHandler("~/")));
+            var route = new Route(
+                "establishment/{urn}/governance",
+                new PageRouteHandler("~/"));
+            var routeData = new RouteData(route, new PageRouteHandler("~/"));
+            routeData.Values["urn"] = estabUrn;
 
-            mockHttpRequestBase.SetupGet(r => r.QueryString)
-                .Returns(HttpUtility.ParseQueryString(string.Empty));
+            controller.ControllerContext = new ControllerContext(
+                controller.ControllerContext.HttpContext,
+                routeData,
+                controller);
 
             var historicEndDate = DateTime.Now.Date.AddDays(-10);
-
             var governor = new GovernorModel
             {
                 Id = gid,
@@ -649,7 +711,9 @@ namespace Edubase.Web.UIUnitTests.Areas.Governors.Controllers.AddOrEdit
 
             mockGovernorsReadService
                 .Setup(s => s.GetEditorDisplayPolicyAsync(
-                    eLookupGovernorRole.Governor, false, It.IsAny<IPrincipal>()))
+                    eLookupGovernorRole.Governor,
+                    false,
+                    It.IsAny<IPrincipal>()))
                 .ReturnsAsync(new GovernorDisplayPolicy());
 
             mockLayoutHelper
@@ -662,6 +726,8 @@ namespace Edubase.Web.UIUnitTests.Areas.Governors.Controllers.AddOrEdit
                     It.IsAny<Action<GroupModel>>()))
                 .Returns(Task.CompletedTask);
 
+            SetupCommonLookupMocks();
+
             // Act
             var result = await controller.AddEditOrReplace(
                 groupUId: null,
@@ -672,7 +738,6 @@ namespace Edubase.Web.UIUnitTests.Areas.Governors.Controllers.AddOrEdit
             // Assert
             var view = Assert.IsType<ViewResult>(result);
             var vm = Assert.IsType<CreateEditGovernorViewModel>(view.Model);
-
             Assert.Equal(CreateEditGovernorViewModel.EditMode.Edit, vm.Mode);
             Assert.True(vm.IsHistoric);
         }
@@ -683,17 +748,20 @@ namespace Edubase.Web.UIUnitTests.Areas.Governors.Controllers.AddOrEdit
             // Arrange
             var estabUrn = 7778;
             var gid = 124;
+            var controller = BuildController();
 
-            mockControllerContext.SetupGet(c => c.RouteData)
-                .Returns(new RouteData(
-                    new Route("establishment/{urn}/governance", new PageRouteHandler("~/")),
-                    new PageRouteHandler("~/")));
+            var route = new Route(
+                "establishment/{urn}/governance",
+                new PageRouteHandler("~/"));
+            var routeData = new RouteData(route, new PageRouteHandler("~/"));
+            routeData.Values["urn"] = estabUrn;
 
-            mockHttpRequestBase.SetupGet(r => r.QueryString)
-                .Returns(HttpUtility.ParseQueryString(string.Empty));
+            controller.ControllerContext = new ControllerContext(
+                controller.ControllerContext.HttpContext,
+                routeData,
+                controller);
 
             var endDateToday = DateTime.Now.Date;
-
             var governor = new GovernorModel
             {
                 Id = gid,
@@ -710,7 +778,9 @@ namespace Edubase.Web.UIUnitTests.Areas.Governors.Controllers.AddOrEdit
 
             mockGovernorsReadService
                 .Setup(s => s.GetEditorDisplayPolicyAsync(
-                    eLookupGovernorRole.Governor, false, It.IsAny<IPrincipal>()))
+                    eLookupGovernorRole.Governor,
+                    false,
+                    It.IsAny<IPrincipal>()))
                 .ReturnsAsync(new GovernorDisplayPolicy());
 
             mockLayoutHelper
@@ -723,6 +793,8 @@ namespace Edubase.Web.UIUnitTests.Areas.Governors.Controllers.AddOrEdit
                     It.IsAny<Action<GroupModel>>()))
                 .Returns(Task.CompletedTask);
 
+            SetupCommonLookupMocks();
+
             // Act
             var result = await controller.AddEditOrReplace(
                 groupUId: null,
@@ -733,7 +805,6 @@ namespace Edubase.Web.UIUnitTests.Areas.Governors.Controllers.AddOrEdit
             // Assert
             var view = Assert.IsType<ViewResult>(result);
             var vm = Assert.IsType<CreateEditGovernorViewModel>(view.Model);
-
             Assert.Equal(CreateEditGovernorViewModel.EditMode.Edit, vm.Mode);
             Assert.False(vm.IsHistoric);
         }
@@ -746,19 +817,26 @@ namespace Edubase.Web.UIUnitTests.Areas.Governors.Controllers.AddOrEdit
             var gid = 200;
             var replacementId = 300;
 
-            mockControllerContext.SetupGet(c => c.RouteData)
-                .Returns(new RouteData(
-                    new Route("establishment/{establishmentUrn}/governance/Replace/{gid}",
-                              new PageRouteHandler("~/")),
-                    new PageRouteHandler("~/")));
-
             var qs = HttpUtility.ParseQueryString(string.Empty);
             qs["gid2"] = replacementId.ToString();
             qs["d"] = "10";
             qs["m"] = "12";
             qs["y"] = "2024";
             qs["rag"] = "true";
-            mockHttpRequestBase.SetupGet(r => r.QueryString).Returns(qs);
+
+            var controller = BuildController(qs);
+
+            var route = new Route(
+                "establishment/{establishmentUrn}/governance/Replace/{gid}",
+                new PageRouteHandler("~/"));
+            var routeData = new RouteData(route, new PageRouteHandler("~/"));
+            routeData.Values["establishmentUrn"] = estabUrn;
+            routeData.Values["gid"] = gid;
+
+            controller.ControllerContext = new ControllerContext(
+                controller.ControllerContext.HttpContext,
+                routeData,
+                controller);
 
             var chair = new GovernorModel
             {
@@ -777,16 +855,16 @@ namespace Edubase.Web.UIUnitTests.Areas.Governors.Controllers.AddOrEdit
             var dto = new GovernorsDetailsDto
             {
                 CurrentGovernors = new List<GovernorModel>
+            {
+                new GovernorModel
                 {
-                    new GovernorModel
-                    {
-                        Id = replacementId,
-                        RoleId = (int)eLookupGovernorRole.Governor,
-                        Person_FirstName = "Beth",
-                        Person_LastName = "Gov",
-                        Person_TitleId = 1
-                    }
+                    Id = replacementId,
+                    RoleId = (int)eLookupGovernorRole.Governor,
+                    Person_FirstName = "Beth",
+                    Person_LastName = "Gov",
+                    Person_TitleId = 1
                 }
+            }
             };
 
             mockGovernorsReadService
@@ -810,6 +888,8 @@ namespace Edubase.Web.UIUnitTests.Areas.Governors.Controllers.AddOrEdit
                     It.IsAny<Action<GroupModel>>()))
                 .Returns(Task.CompletedTask);
 
+            SetupCommonLookupMocks();
+
             // Act
             var result = await controller.AddEditOrReplace(
                 groupUId: null,
@@ -820,11 +900,8 @@ namespace Edubase.Web.UIUnitTests.Areas.Governors.Controllers.AddOrEdit
             // Assert
             var view = Assert.IsType<ViewResult>(result);
             var vm = Assert.IsType<CreateEditGovernorViewModel>(view.Model);
-
             Assert.Equal(CreateEditGovernorViewModel.EditMode.Replace, vm.Mode);
-
             Assert.True(vm.ReinstateAsGovernor);
-
             Assert.Equal(10, vm.ReplaceGovernorViewModel.AppointmentEndDate.Day);
             Assert.Equal(12, vm.ReplaceGovernorViewModel.AppointmentEndDate.Month);
             Assert.Equal(2024, vm.ReplaceGovernorViewModel.AppointmentEndDate.Year);
@@ -833,39 +910,59 @@ namespace Edubase.Web.UIUnitTests.Areas.Governors.Controllers.AddOrEdit
         /// <summary>
         /// Every chair of local governing body role, combined with every chair of local governing body role.
         /// At the time of writing, equivalent to:
-        ///     [InlineData( eLookupGovernorRole.Group_SharedChairOfLocalGoverningBody,          eLookupGovernorRole.Group_SharedChairOfLocalGoverningBody          )]
-        ///     [InlineData( eLookupGovernorRole.Group_SharedChairOfLocalGoverningBody,          eLookupGovernorRole.Establishment_SharedChairOfLocalGoverningBody  )]
-        ///     [InlineData( eLookupGovernorRole.Establishment_SharedChairOfLocalGoverningBody,  eLookupGovernorRole.Group_SharedChairOfLocalGoverningBody          )]
-        ///     [InlineData( eLookupGovernorRole.Establishment_SharedChairOfLocalGoverningBody,  eLookupGovernorRole.Establishment_SharedChairOfLocalGoverningBody  )]
+        /// [InlineData( eLookupGovernorRole.Group_SharedChairOfLocalGoverningBody, eLookupGovernorRole.Group_SharedChairOfLocalGoverningBody )]
+        /// [InlineData( eLookupGovernorRole.Group_SharedChairOfLocalGoverningBody, eLookupGovernorRole.Establishment_SharedChairOfLocalGoverningBody )]
+        /// [InlineData( eLookupGovernorRole.Establishment_SharedChairOfLocalGoverningBody, eLookupGovernorRole.Group_SharedChairOfLocalGoverningBody )]
+        /// [InlineData( eLookupGovernorRole.Establishment_SharedChairOfLocalGoverningBody, eLookupGovernorRole.Establishment_SharedChairOfLocalGoverningBody )]
         /// </summary>
         public static IEnumerable<object[]> PairwiseChairOfLocalGoverningBodyRoles =>
             from preExistingRole in EnumSets.eChairOfLocalGoverningBodyRoles
             from newRole in EnumSets.eChairOfLocalGoverningBodyRoles
             select new object[] { preExistingRole, newRole };
 
-        [Theory()]
+        [Theory]
         [MemberData(nameof(PairwiseChairOfLocalGoverningBodyRoles))]
-        public async Task Gov_AddEditOrReplace_RoleSpecified_ChairOfLocalGoverningBody_RoleAlreadyExists(eLookupGovernorRole preExistingGovernorRole, eLookupGovernorRole newGovernorRole)
+        public async Task Gov_AddEditOrReplace_RoleSpecified_ChairOfLocalGoverningBody_RoleAlreadyExists(
+            eLookupGovernorRole preExistingGovernorRole,
+            eLookupGovernorRole newGovernorRole)
         {
+            // Arrange
             var estabUrn = 4;
+            var controller = BuildController();
+
             mockGovernorsReadService
                 .Setup(g => g.GetGovernorListAsync(estabUrn, null, It.IsAny<IPrincipal>()))
                 .ReturnsAsync(() => new GovernorsDetailsDto
                 {
-                    CurrentGovernors = new List<GovernorModel>()
+                    CurrentGovernors = new List<GovernorModel>
                     {
-                        new GovernorModel() {RoleId = (int) preExistingGovernorRole}
+                    new GovernorModel { RoleId = (int)preExistingGovernorRole }
                     }
                 });
-            mockControllerContext.SetupGet(c => c.RouteData)
-                .Returns(new RouteData(new Route("", new PageRouteHandler("~/")), new PageRouteHandler("~/")));
 
-            var result = await controller.AddEditOrReplace(null, estabUrn, newGovernorRole, null);
+            var route = new Route("", new PageRouteHandler("~/"));
+            var routeData = new RouteData(route, new PageRouteHandler("~/"));
+            routeData.Values["controller"] = "Governors";
+            routeData.Values["action"] = "AddEditOrReplace";
 
+            controller.ControllerContext = new ControllerContext(
+                controller.ControllerContext.HttpContext,
+                routeData,
+                controller);
+
+            SetupCommonLookupMocks();
+
+            // Act
+            var result = await controller.AddEditOrReplace(
+                groupUId: null,
+                establishmentUrn: estabUrn,
+                role: newGovernorRole,
+                gid: null);
+
+            // Assert
             var redirectResult = Assert.IsType<RedirectToRouteResult>(result);
 
             string expectedRoute;
-
             if (newGovernorRole == eLookupGovernorRole.Group_SharedChairOfLocalGoverningBody)
             {
                 expectedRoute = "SelectSharedGovernor";
@@ -878,7 +975,6 @@ namespace Edubase.Web.UIUnitTests.Areas.Governors.Controllers.AddOrEdit
             Assert.Equal(expectedRoute, redirectResult.RouteName);
         }
 
-
         // Delegate to EnumSets
         public static TheoryData<eLookupGovernorRole, eLookupGovernorRole> ForbiddenCombinationsOfGovernanceProfessionalRoles
         {
@@ -889,19 +985,22 @@ namespace Edubase.Web.UIUnitTests.Areas.Governors.Controllers.AddOrEdit
                 {
                     theoryData.Add((eLookupGovernorRole) combination[0], (eLookupGovernorRole) combination[1]);
                 }
+
                 return theoryData;
             }
         }
 
-        [Theory()]
+        [Theory]
         [MemberData(nameof(ForbiddenCombinationsOfGovernanceProfessionalRoles))]
-        public async Task Gov_AddEditOrReplace_RoleSpecified_GovernanceProfessional_RoleAlreadyExists_DisallowedThereforeReject(eLookupGovernorRole preExistingGovernorRole, eLookupGovernorRole newGovernorRole)
+        public async Task Gov_AddEditOrReplace_RoleSpecified_GovernanceProfessional_RoleAlreadyExists_DisallowedThereforeReject(
+            eLookupGovernorRole preExistingGovernorRole,
+            eLookupGovernorRole newGovernorRole)
         {
-
+            // Arrange
             var currentGovernors = new List<GovernorModel>
-            {
-                new GovernorModel { RoleId = (int)preExistingGovernorRole }
-            };
+        {
+            new GovernorModel { RoleId = (int)preExistingGovernorRole }
+        };
 
             var governorsDetails = new GovernorsDetailsDto
             {
@@ -911,19 +1010,33 @@ namespace Edubase.Web.UIUnitTests.Areas.Governors.Controllers.AddOrEdit
                 HasFullAccess = true
             };
 
-            mockGovernorsReadService.Setup(g => g.GetGovernorListAsync(It.IsAny<int?>(), It.IsAny<int?>(), It.IsAny<IPrincipal>()))
+            mockGovernorsReadService
+                .Setup(g => g.GetGovernorListAsync(
+                    It.IsAny<int?>(),
+                    It.IsAny<int?>(),
+                    It.IsAny<IPrincipal>()))
                 .ReturnsAsync(governorsDetails);
 
-            var result = await controller.RoleAllowed(newGovernorRole, null, null, null, false);
+            var controller = BuildController();
 
+            // Act
+            var result = await controller.RoleAllowed(
+                newGovernorRole,
+                null,
+                null,
+                null,
+                false);
+
+            // Assert
             Assert.False(result);
         }
 
-        [Fact()]
+        [Fact]
         public async Task Gov_AddEditOrReplace_Create_Defaults_OriginalChairOfTrustees_ToNo()
         {
             // Arrange
             var groupId = 42;
+            var controller = BuildController();
 
             var governorDetailsDto = new GovernorsDetailsDto
             {
@@ -931,20 +1044,25 @@ namespace Edubase.Web.UIUnitTests.Areas.Governors.Controllers.AddOrEdit
                 HistoricalGovernors = new List<GovernorModel>(),
                 ApplicableRoles = new List<eLookupGovernorRole> { eLookupGovernorRole.ChairOfTrustees },
                 RoleDisplayPolicies = new Dictionary<eLookupGovernorRole, GovernorDisplayPolicy>
-                {
-                    { eLookupGovernorRole.ChairOfTrustees, new GovernorDisplayPolicy() }
-                }
-            };
+            {
+                { eLookupGovernorRole.ChairOfTrustees, new GovernorDisplayPolicy() }
+            }};
 
             mockGovernorsReadService
                 .Setup(g => g.GetGovernorListAsync(null, groupId, It.IsAny<IPrincipal>()))
                 .ReturnsAsync(() => governorDetailsDto);
 
-            mockControllerContext
-                .SetupGet(c => c.RouteData)
-                .Returns(new RouteData(
-                    new Route("", new PageRouteHandler("~/")),
-                    new PageRouteHandler("~/")));
+            var route = new Route("", new PageRouteHandler("~/"));
+            var routeData = new RouteData(route, new PageRouteHandler("~/"));
+            routeData.Values["controller"] = "Governors";
+            routeData.Values["action"] = "AddEditOrReplace";
+
+            controller.ControllerContext = new ControllerContext(
+                    controller.ControllerContext.HttpContext,
+                    routeData,
+                    controller);
+
+            SetupCommonLookupMocks();
 
             mockLayoutHelper
                 .Setup(l => l.PopulateLayoutProperties(
@@ -964,12 +1082,15 @@ namespace Edubase.Web.UIUnitTests.Areas.Governors.Controllers.AddOrEdit
                 .ReturnsAsync(() => new GovernorDisplayPolicy());
 
             // Act
-            var result = await controller.AddEditOrReplace(groupId, null, eLookupGovernorRole.ChairOfTrustees, null);
+            var result = await controller.AddEditOrReplace(
+                groupUId: groupId,
+                establishmentUrn: null,
+                role: eLookupGovernorRole.ChairOfTrustees,
+                gid: null);
 
             // Assert
             var viewResult = Assert.IsType<ViewResult>(result);
             var model = Assert.IsType<CreateEditGovernorViewModel>(viewResult.Model);
-
             Assert.Null(model.IsOriginalChairOfTrustees);
 
             var noItem = Assert.Single(model.YesNoSelect.Where(i => i.Value == "false"));
@@ -982,6 +1103,19 @@ namespace Edubase.Web.UIUnitTests.Areas.Governors.Controllers.AddOrEdit
             // Arrange
             var estabUrn = 4000;
             var governorId = 200;
+            var controller = BuildController();
+
+            var route = new Route("", new PageRouteHandler("~/"));
+            var routeData = new RouteData(route, new PageRouteHandler("~/"));
+            routeData.Values["controller"] = "Governors";
+            routeData.Values["action"] = "AddEditOrReplace";
+
+            controller.ControllerContext = new ControllerContext(
+                    controller.ControllerContext.HttpContext,
+                    routeData,
+                    controller);
+
+            SetupCommonLookupMocks();
 
             var governor = new GovernorModel
             {
@@ -990,10 +1124,6 @@ namespace Edubase.Web.UIUnitTests.Areas.Governors.Controllers.AddOrEdit
                 IsOriginalSignatoryMember = true,
                 IsOriginalChairOfTrustees = true
             };
-
-            mockControllerContext
-                .SetupGet(c => c.RouteData)
-                .Returns(new RouteData(new Route("", new PageRouteHandler("~/")), new PageRouteHandler("~/")));
 
             mockGovernorsReadService
                 .Setup(g => g.GetGovernorAsync(governorId, It.IsAny<IPrincipal>()))
@@ -1020,7 +1150,11 @@ namespace Edubase.Web.UIUnitTests.Areas.Governors.Controllers.AddOrEdit
                 .Returns(Task.CompletedTask);
 
             // Act
-            var result = await controller.AddEditOrReplace(null, estabUrn, null, governorId);
+            var result = await controller.AddEditOrReplace(
+                groupUId: null,
+                establishmentUrn: estabUrn,
+                role: null,
+                gid: governorId);
 
             // Assert
             var viewResult = Assert.IsType<ViewResult>(result);
@@ -1028,8 +1162,7 @@ namespace Edubase.Web.UIUnitTests.Areas.Governors.Controllers.AddOrEdit
 
             Assert.True(vm.IsOriginalSignatoryMember);
             Assert.True(vm.IsOriginalChairOfTrustees);
-
-            // Do NOT assert vm.DisplayPolicy.AppointingBodyId (frontend-only rule - JS)
+            // AppointingBodyId is intentionally not asserted (frontend-only rule via JS)
         }
 
         [Fact]
@@ -1037,12 +1170,12 @@ namespace Edubase.Web.UIUnitTests.Areas.Governors.Controllers.AddOrEdit
         {
             // Arrange
             var estabUrn = 1010;
+            var controller = BuildController();
 
             var vm = new CreateEditGovernorViewModel
             {
                 Mode = CreateEditGovernorViewModel.EditMode.Edit,
                 GovernorRole = eLookupGovernorRole.Governor,
-
                 AppointingBodyId = 5,
                 AppointmentStartDate = new DateTimeViewModel(new DateTime(2020, 1, 1)),
                 AppointmentEndDate = new DateTimeViewModel(new DateTime(2025, 1, 1)),
@@ -1051,27 +1184,25 @@ namespace Edubase.Web.UIUnitTests.Areas.Governors.Controllers.AddOrEdit
                 GroupUId = null,
                 EstablishmentUrn = estabUrn,
                 GID = 444,
-
                 FirstName = "Jane",
                 MiddleName = "Elizabeth",
                 LastName = "Doe",
-
                 IsOriginalChairOfTrustees = true,
                 IsOriginalSignatoryMember = false,
-
                 GovernorTitleId = 3,
-
                 PreviousFirstName = "J.",
                 PreviousMiddleName = "E.",
                 PreviousLastName = "D.",
                 PreviousTitleId = 7,
-
                 PostCode = "XY1 2ZZ",
                 TelephoneNumber = "0800123456"
             };
 
             mockGovernorsReadService
-                .Setup(g => g.GetEditorDisplayPolicyAsync(It.IsAny<eLookupGovernorRole>(), It.IsAny<bool>(), It.IsAny<IPrincipal>()))
+                .Setup(g => g.GetEditorDisplayPolicyAsync(
+                    It.IsAny<eLookupGovernorRole>(),
+                    It.IsAny<bool>(),
+                    It.IsAny<IPrincipal>()))
                 .ReturnsAsync(new GovernorDisplayPolicy());
 
             mockGovernorsWriteService
@@ -1089,20 +1220,16 @@ namespace Edubase.Web.UIUnitTests.Areas.Governors.Controllers.AddOrEdit
                     Assert.Equal("person@example.com", gm.EmailAddress);
                     Assert.Equal(estabUrn, gm.EstablishmentUrn);
                     Assert.Equal(444, gm.Id);
-
                     Assert.Equal("Jane", gm.Person_FirstName);
                     Assert.Equal("Elizabeth", gm.Person_MiddleName);
                     Assert.Equal("Doe", gm.Person_LastName);
-
                     Assert.True(gm.IsOriginalChairOfTrustees);
                     Assert.False(gm.IsOriginalSignatoryMember);
                     Assert.Equal(3, gm.Person_TitleId);
-
                     Assert.Equal("J.", gm.PreviousPerson_FirstName);
                     Assert.Equal("E.", gm.PreviousPerson_MiddleName);
                     Assert.Equal("D.", gm.PreviousPerson_LastName);
                     Assert.Equal(7, gm.PreviousPerson_TitleId);
-
                     Assert.Equal("XY1 2ZZ", gm.PostCode);
                     Assert.Equal((int) eLookupGovernorRole.Governor, gm.RoleId);
                     Assert.Equal("0800123456", gm.TelephoneNumber);
@@ -1119,8 +1246,13 @@ namespace Edubase.Web.UIUnitTests.Areas.Governors.Controllers.AddOrEdit
                     It.IsAny<Action<GroupModel>>()))
                 .Returns(Task.CompletedTask);
 
+            SetupCommonLookupMocks();
+
             // Act
             await controller.AddEditOrReplace(vm);
+
+            // Assert
+            // Assertions are inside the SaveAsync callback
         }
 
         [Fact]
@@ -1128,29 +1260,25 @@ namespace Edubase.Web.UIUnitTests.Areas.Governors.Controllers.AddOrEdit
         {
             // Arrange
             var estabUrn = 2020;
-
             var endDate = new DateTime(2024, 6, 15);
+            var controller = BuildController();
 
             var vm = new CreateEditGovernorViewModel
             {
                 Mode = CreateEditGovernorViewModel.EditMode.Replace,
                 GovernorRole = eLookupGovernorRole.Governor,
-
                 ReplaceGovernorViewModel = new ReplaceGovernorViewModel
                 {
                     AppointmentEndDate = new DateTimeViewModel(endDate)
                 },
-
                 FirstName = "Mark",
                 MiddleName = "Alan",
                 LastName = "River",
-
                 GovernorTitleId = 2,
                 PreviousFirstName = "M.",
                 PreviousMiddleName = "A.",
                 PreviousLastName = "R.",
                 PreviousTitleId = 9,
-
                 AppointingBodyId = 10,
                 AppointmentEndDate = new DateTimeViewModel(new DateTime(2025, 1, 1)),
                 DOB = new DateTimeViewModel(new DateTime(1985, 4, 4)),
@@ -1161,7 +1289,10 @@ namespace Edubase.Web.UIUnitTests.Areas.Governors.Controllers.AddOrEdit
             };
 
             mockGovernorsReadService
-                .Setup(g => g.GetEditorDisplayPolicyAsync(It.IsAny<eLookupGovernorRole>(), It.IsAny<bool>(), It.IsAny<IPrincipal>()))
+                .Setup(g => g.GetEditorDisplayPolicyAsync(
+                    It.IsAny<eLookupGovernorRole>(),
+                    It.IsAny<bool>(),
+                    It.IsAny<IPrincipal>()))
                 .ReturnsAsync(new GovernorDisplayPolicy());
 
             mockGovernorsWriteService
@@ -1173,18 +1304,14 @@ namespace Edubase.Web.UIUnitTests.Areas.Governors.Controllers.AddOrEdit
                 .Callback<GovernorModel, IPrincipal>((gm, _) =>
                 {
                     Assert.Equal(endDate.AddDays(1), gm.AppointmentStartDate);
-
                     Assert.Equal("Mark", gm.Person_FirstName);
                     Assert.Equal("Alan", gm.Person_MiddleName);
                     Assert.Equal("River", gm.Person_LastName);
-
                     Assert.Equal(2, gm.Person_TitleId);
-
                     Assert.Equal("M.", gm.PreviousPerson_FirstName);
                     Assert.Equal("A.", gm.PreviousPerson_MiddleName);
                     Assert.Equal("R.", gm.PreviousPerson_LastName);
                     Assert.Equal(9, gm.PreviousPerson_TitleId);
-
                     Assert.Equal(10, gm.AppointingBodyId);
                     Assert.Equal(estabUrn, gm.EstablishmentUrn);
                     Assert.Equal("mark@example.com", gm.EmailAddress);
@@ -1203,8 +1330,13 @@ namespace Edubase.Web.UIUnitTests.Areas.Governors.Controllers.AddOrEdit
                     It.IsAny<Action<GroupModel>>()))
                 .Returns(Task.CompletedTask);
 
+            SetupCommonLookupMocks();
+
             // Act
             await controller.AddEditOrReplace(vm);
+
+            // Assert
+            // Assertions are inside the SaveAsync callback
         }
 
         [Fact]
@@ -1212,11 +1344,12 @@ namespace Edubase.Web.UIUnitTests.Areas.Governors.Controllers.AddOrEdit
         {
             // Arrange
             var groupId = 900;
+            var controller = BuildController();
 
             var vm = new CreateEditGovernorViewModel
             {
                 GroupUId = groupId,
-                GovernorRole = eLookupGovernorRole.Establishment_SharedLocalGovernor, // shared role
+                GovernorRole = eLookupGovernorRole.Establishment_SharedLocalGovernor,
                 FirstName = "Alex",
                 MiddleName = "J",
                 LastName = "Taylor",
@@ -1235,6 +1368,15 @@ namespace Edubase.Web.UIUnitTests.Areas.Governors.Controllers.AddOrEdit
                 RoleId = (int) eLookupGovernorRole.Establishment_SharedLocalGovernor
             };
 
+            var route = new Route("", new PageRouteHandler("~/"));
+            var rd = new RouteData(route, new PageRouteHandler("~/"));
+            controller.ControllerContext = new ControllerContext(
+                controller.ControllerContext.HttpContext,
+                rd,
+                controller);
+
+            SetupCommonLookupMocks();
+
             mockGovernorsReadService
                 .Setup(g => g.GetGovernorListAsync(null, groupId, It.IsAny<IPrincipal>()))
                 .ReturnsAsync(new GovernorsDetailsDto
@@ -1243,11 +1385,11 @@ namespace Edubase.Web.UIUnitTests.Areas.Governors.Controllers.AddOrEdit
                     HistoricalGovernors = new List<GovernorModel>(),
                     ApplicableRoles = new List<eLookupGovernorRole>
                     {
-                eLookupGovernorRole.Establishment_SharedLocalGovernor
+                    eLookupGovernorRole.Establishment_SharedLocalGovernor
                     },
                     RoleDisplayPolicies = new Dictionary<eLookupGovernorRole, GovernorDisplayPolicy>
                     {
-                { eLookupGovernorRole.Establishment_SharedLocalGovernor, new GovernorDisplayPolicy() }
+                    { eLookupGovernorRole.Establishment_SharedLocalGovernor, new GovernorDisplayPolicy() }
                     }
                 });
 
@@ -1278,7 +1420,6 @@ namespace Edubase.Web.UIUnitTests.Areas.Governors.Controllers.AddOrEdit
             // Assert
             var redirect = Assert.IsType<RedirectToRouteResult>(result);
             Assert.Equal("GroupEditGovernance", redirect.RouteName);
-
             Assert.Equal(groupId, redirect.RouteValues["groupUId"]);
             Assert.Equal(existing.Id, redirect.RouteValues["duplicateGovernorId"]);
 
@@ -1292,6 +1433,7 @@ namespace Edubase.Web.UIUnitTests.Areas.Governors.Controllers.AddOrEdit
         {
             // Arrange
             var groupId = 901;
+            var controller = BuildController();
 
             var vm = new CreateEditGovernorViewModel
             {
@@ -1317,34 +1459,34 @@ namespace Edubase.Web.UIUnitTests.Areas.Governors.Controllers.AddOrEdit
 
             mockGovernorsReadService
                 .Setup(g => g.GetGovernorListAsync(null, groupId, It.IsAny<IPrincipal>()))
-                .Returns(Task.FromResult(new GovernorsDetailsDto
+                .ReturnsAsync(new GovernorsDetailsDto
                 {
                     CurrentGovernors = new List<GovernorModel> { existing },
                     HistoricalGovernors = new List<GovernorModel>(),
                     ApplicableRoles = new List<eLookupGovernorRole>
                     {
-                        eLookupGovernorRole.Establishment_SharedLocalGovernor
+                    eLookupGovernorRole.Establishment_SharedLocalGovernor
                     },
                     RoleDisplayPolicies = new Dictionary<eLookupGovernorRole, GovernorDisplayPolicy>
                     {
-                        { eLookupGovernorRole.Establishment_SharedLocalGovernor, new GovernorDisplayPolicy() }
+                    { eLookupGovernorRole.Establishment_SharedLocalGovernor, new GovernorDisplayPolicy() }
                     }
-                }));
+                });
 
             mockGovernorsReadService
                 .Setup(g => g.GetEditorDisplayPolicyAsync(
                     eLookupGovernorRole.Establishment_SharedLocalGovernor,
                     true,
                     It.IsAny<IPrincipal>()))
-                .Returns(Task.FromResult(new GovernorDisplayPolicy()));
+                .ReturnsAsync(new GovernorDisplayPolicy());
 
             mockGovernorsWriteService
                 .Setup(g => g.ValidateAsync(It.IsAny<GovernorModel>(), It.IsAny<IPrincipal>()))
-                .Returns(Task.FromResult(new ValidationEnvelopeDto()));
+                .ReturnsAsync(new ValidationEnvelopeDto());
 
             mockGovernorsWriteService
                 .Setup(g => g.SaveAsync(It.IsAny<GovernorModel>(), It.IsAny<IPrincipal>()))
-                .Returns(Task.FromResult(new ApiResponse<int>(true) { Response = 999 }));
+                .ReturnsAsync(new ApiResponse<int>(true) { Response = 999 });
 
             mockLayoutHelper
                 .Setup(l => l.PopulateLayoutProperties(
@@ -1355,6 +1497,8 @@ namespace Edubase.Web.UIUnitTests.Areas.Governors.Controllers.AddOrEdit
                     It.IsAny<Action<EstablishmentModel>>(),
                     It.IsAny<Action<GroupModel>>()))
                 .Returns(Task.CompletedTask);
+
+            SetupCommonLookupMocks();
 
             // Act
             var result = await controller.AddEditOrReplace(vm);
@@ -1373,6 +1517,7 @@ namespace Edubase.Web.UIUnitTests.Areas.Governors.Controllers.AddOrEdit
         {
             // Arrange
             var groupId = 902;
+            var controller = BuildController();
 
             var vm = new CreateEditGovernorViewModel
             {
@@ -1398,34 +1543,34 @@ namespace Edubase.Web.UIUnitTests.Areas.Governors.Controllers.AddOrEdit
 
             mockGovernorsReadService
                 .Setup(g => g.GetGovernorListAsync(null, groupId, It.IsAny<IPrincipal>()))
-                .Returns(Task.FromResult(new GovernorsDetailsDto
+                .ReturnsAsync(new GovernorsDetailsDto
                 {
                     CurrentGovernors = new List<GovernorModel> { existing },
                     HistoricalGovernors = new List<GovernorModel>(),
                     ApplicableRoles = new List<eLookupGovernorRole>
                     {
-                        eLookupGovernorRole.Establishment_SharedLocalGovernor
+                    eLookupGovernorRole.Establishment_SharedLocalGovernor
                     },
                     RoleDisplayPolicies = new Dictionary<eLookupGovernorRole, GovernorDisplayPolicy>
                     {
-                        { eLookupGovernorRole.Establishment_SharedLocalGovernor, new GovernorDisplayPolicy() }
+                    { eLookupGovernorRole.Establishment_SharedLocalGovernor, new GovernorDisplayPolicy() }
                     }
-                }));
+                });
 
             mockGovernorsReadService
                 .Setup(g => g.GetEditorDisplayPolicyAsync(
                     eLookupGovernorRole.Establishment_SharedLocalGovernor,
                     true,
                     It.IsAny<IPrincipal>()))
-                .Returns(Task.FromResult(new GovernorDisplayPolicy()));
+                .ReturnsAsync(new GovernorDisplayPolicy());
 
             mockGovernorsWriteService
                 .Setup(g => g.ValidateAsync(It.IsAny<GovernorModel>(), It.IsAny<IPrincipal>()))
-                .Returns(Task.FromResult(new ValidationEnvelopeDto()));
+                .ReturnsAsync(new ValidationEnvelopeDto());
 
             mockGovernorsWriteService
                 .Setup(g => g.SaveAsync(It.IsAny<GovernorModel>(), It.IsAny<IPrincipal>()))
-                .Returns(Task.FromResult(new ApiResponse<int>(true) { Response = 1000 }));
+                .ReturnsAsync(new ApiResponse<int>(true) { Response = 1000 });
 
             mockLayoutHelper
                 .Setup(l => l.PopulateLayoutProperties(
@@ -1436,6 +1581,8 @@ namespace Edubase.Web.UIUnitTests.Areas.Governors.Controllers.AddOrEdit
                     It.IsAny<Action<EstablishmentModel>>(),
                     It.IsAny<Action<GroupModel>>()))
                 .Returns(Task.CompletedTask);
+
+            SetupCommonLookupMocks();
 
             // Act
             var result = await controller.AddEditOrReplace(vm);
@@ -1454,6 +1601,7 @@ namespace Edubase.Web.UIUnitTests.Areas.Governors.Controllers.AddOrEdit
         {
             // Arrange
             var estabUrn = 1234;
+            var controller = BuildController();
 
             var vm = new CreateEditGovernorViewModel
             {
@@ -1467,20 +1615,26 @@ namespace Edubase.Web.UIUnitTests.Areas.Governors.Controllers.AddOrEdit
                 GovernorTitleId = 1
             };
 
-            // Display policy needed early in POST
+            var route = new Route("", new PageRouteHandler("~/"));
+            var rd = new RouteData(route, new PageRouteHandler("~/"));
+            controller.ControllerContext = new ControllerContext(
+                controller.ControllerContext.HttpContext,
+                rd,
+                controller);
+
+            SetupCommonLookupMocks();
+
             mockGovernorsReadService
                 .Setup(g => g.GetEditorDisplayPolicyAsync(
                     eLookupGovernorRole.Governor,
-                    false,          // groupUId.HasValue == false
+                    false,
                     It.IsAny<IPrincipal>()))
-                .Returns(Task.FromResult(new GovernorDisplayPolicy()));
+                .ReturnsAsync(new GovernorDisplayPolicy());
 
-            // Validation passes – so we enter the ModelState.IsValid block
             mockGovernorsWriteService
                 .Setup(g => g.ValidateAsync(It.IsAny<GovernorModel>(), It.IsAny<IPrincipal>()))
-                .Returns(Task.FromResult(new ValidationEnvelopeDto())); // no errors
+                .ReturnsAsync(new ValidationEnvelopeDto());
 
-            // Save fails – triggers ErrorsToModelState
             var apiError = new ApiError
             {
                 Fields = "Person_FirstName",
@@ -1489,10 +1643,10 @@ namespace Edubase.Web.UIUnitTests.Areas.Governors.Controllers.AddOrEdit
 
             mockGovernorsWriteService
                 .Setup(g => g.SaveAsync(It.IsAny<GovernorModel>(), It.IsAny<IPrincipal>()))
-                .Returns(Task.FromResult(new ApiResponse<int>(false)
+                .ReturnsAsync(new ApiResponse<int>(false)
                 {
                     Errors = new[] { apiError }
-                }));
+                });
 
             mockLayoutHelper
                 .Setup(l => l.PopulateLayoutProperties(
@@ -1529,6 +1683,7 @@ namespace Edubase.Web.UIUnitTests.Areas.Governors.Controllers.AddOrEdit
         {
             // Arrange
             var estabUrn = 5678;
+            var controller = BuildController();
 
             var vm = new CreateEditGovernorViewModel
             {
@@ -1547,15 +1702,15 @@ namespace Edubase.Web.UIUnitTests.Areas.Governors.Controllers.AddOrEdit
                     eLookupGovernorRole.Governor,
                     false,
                     It.IsAny<IPrincipal>()))
-                .Returns(Task.FromResult(new GovernorDisplayPolicy()));
+                .ReturnsAsync(new GovernorDisplayPolicy());
 
             mockGovernorsWriteService
                 .Setup(g => g.ValidateAsync(It.IsAny<GovernorModel>(), It.IsAny<IPrincipal>()))
-                .Returns(Task.FromResult(new ValidationEnvelopeDto())); // no errors
+                .ReturnsAsync(new ValidationEnvelopeDto());
 
             mockGovernorsWriteService
                 .Setup(g => g.SaveAsync(It.IsAny<GovernorModel>(), It.IsAny<IPrincipal>()))
-                .Returns(Task.FromResult(new ApiResponse<int>(true) { Response = 888 }));
+                .ReturnsAsync(new ApiResponse<int>(true) { Response = 888 });
 
             mockLayoutHelper
                 .Setup(l => l.PopulateLayoutProperties(
@@ -1567,12 +1722,13 @@ namespace Edubase.Web.UIUnitTests.Areas.Governors.Controllers.AddOrEdit
                     It.IsAny<Action<GroupModel>>()))
                 .Returns(Task.CompletedTask);
 
+            SetupCommonLookupMocks();
+
             // Act
             var result = await controller.AddEditOrReplace(vm);
 
             // Assert
             var redirect = Assert.IsType<RedirectResult>(result);
-
             Assert.EndsWith("#school-governance", redirect.Url);
 
             mockGovernorsWriteService.Verify(
@@ -1586,6 +1742,7 @@ namespace Edubase.Web.UIUnitTests.Areas.Governors.Controllers.AddOrEdit
             // Arrange
             var estabUrn = 5000;
             var oldGovernorId = 321;
+            var controller = BuildController();
 
             var vm = new CreateEditGovernorViewModel
             {
@@ -1609,19 +1766,19 @@ namespace Edubase.Web.UIUnitTests.Areas.Governors.Controllers.AddOrEdit
                     eLookupGovernorRole.Governor,
                     false,
                     It.IsAny<IPrincipal>()))
-                .Returns(Task.FromResult(new GovernorDisplayPolicy()));
+                .ReturnsAsync(new GovernorDisplayPolicy());
 
             mockGovernorsWriteService
                 .Setup(g => g.ValidateAsync(It.IsAny<GovernorModel>(), It.IsAny<IPrincipal>()))
-                .Returns(Task.FromResult(new ValidationEnvelopeDto()));
+                .ReturnsAsync(new ValidationEnvelopeDto());
 
             mockGovernorsReadService
                 .Setup(g => g.GetGovernorAsync(oldGovernorId, It.IsAny<IPrincipal>()))
-                .Returns(Task.FromResult(new GovernorModel { Id = oldGovernorId }));
+                .ReturnsAsync(new GovernorModel { Id = oldGovernorId });
 
             mockGovernorsWriteService
                 .Setup(g => g.SaveAsync(It.IsAny<GovernorModel>(), It.IsAny<IPrincipal>()))
-                .Returns(Task.FromResult(new ApiResponse<int>(true) { Response = 999 }));
+                .ReturnsAsync(new ApiResponse<int>(true) { Response = 999 });
 
             mockLayoutHelper
                 .Setup(l => l.PopulateLayoutProperties(
@@ -1632,6 +1789,9 @@ namespace Edubase.Web.UIUnitTests.Areas.Governors.Controllers.AddOrEdit
                     It.IsAny<Action<EstablishmentModel>>(),
                     It.IsAny<Action<GroupModel>>()))
                 .Returns(Task.CompletedTask);
+
+
+            SetupCommonLookupMocks();
 
             // Act
             var result = await controller.AddEditOrReplace(vm);
@@ -1651,13 +1811,14 @@ namespace Edubase.Web.UIUnitTests.Areas.Governors.Controllers.AddOrEdit
             var estabUrn = 7000;
             var previousId = 111;
             var replacementEndDate = new DateTime(2025, 6, 1);
+            var controller = BuildController();
 
             var vm = new CreateEditGovernorViewModel
             {
                 GovernorRole = eLookupGovernorRole.Governor,
                 EstablishmentUrn = estabUrn,
-                SelectedPreviousGovernorId = previousId,          // REQUIRED
-                ReinstateAsGovernor = false,                      // IMPORTANT for this test
+                SelectedPreviousGovernorId = previousId,
+                ReinstateAsGovernor = false,
                 ReplaceGovernorViewModel = new ReplaceGovernorViewModel
                 {
                     AppointmentEndDate = new DateTimeViewModel(replacementEndDate)
@@ -1675,21 +1836,19 @@ namespace Edubase.Web.UIUnitTests.Areas.Governors.Controllers.AddOrEdit
                     eLookupGovernorRole.Governor,
                     false,
                     It.IsAny<IPrincipal>()))
-                .Returns(Task.FromResult(new GovernorDisplayPolicy()));
+                .ReturnsAsync(new GovernorDisplayPolicy());
 
-            // Validation passes
             mockGovernorsWriteService
                 .Setup(g => g.ValidateAsync(It.IsAny<GovernorModel>(), It.IsAny<IPrincipal>()))
-                .Returns(Task.FromResult(new ValidationEnvelopeDto()));
+                .ReturnsAsync(new ValidationEnvelopeDto());
 
-            // Main save succeeds
             mockGovernorsWriteService
                 .Setup(g => g.SaveAsync(It.IsAny<GovernorModel>(), It.IsAny<IPrincipal>()))
-                .Returns(Task.FromResult(new ApiResponse<int>(true) { Response = 999 }));
+                .ReturnsAsync(new ApiResponse<int>(true) { Response = 999 });
 
             mockGovernorsWriteService
                 .Setup(g => g.UpdateDatesAsync(It.IsAny<int>(), It.IsAny<DateTime>(), It.IsAny<IPrincipal>()))
-                .Returns(Task.FromResult(new ApiResponse(true)));
+                .ReturnsAsync(new ApiResponse(true));
 
             mockLayoutHelper
                 .Setup(l => l.PopulateLayoutProperties(
@@ -1701,16 +1860,16 @@ namespace Edubase.Web.UIUnitTests.Areas.Governors.Controllers.AddOrEdit
                     It.IsAny<Action<GroupModel>>()))
                 .Returns(Task.CompletedTask);
 
+            SetupCommonLookupMocks();
+
             // Act
             var result = await controller.AddEditOrReplace(vm);
 
             // Assert
-            // We should have retired the previous governor with the replacement end date
             mockGovernorsWriteService.Verify(
                 g => g.UpdateDatesAsync(previousId, replacementEndDate, It.IsAny<IPrincipal>()),
                 Times.Once);
 
-            // No extra save for ReInstateChairAsNonChairAsync (only main SaveAsync)
             mockGovernorsWriteService.Verify(
                 g => g.SaveAsync(It.IsAny<GovernorModel>(), It.IsAny<IPrincipal>()),
                 Times.Once);
@@ -1727,6 +1886,7 @@ namespace Edubase.Web.UIUnitTests.Areas.Governors.Controllers.AddOrEdit
             var oldChairId = 250;
             var oldChairEndDate = new DateTime(2024, 3, 1);
             var replacementEndDate = new DateTime(2025, 7, 1);
+            var controller = BuildController();
 
             var vm = new CreateEditGovernorViewModel
             {
@@ -1752,28 +1912,28 @@ namespace Edubase.Web.UIUnitTests.Areas.Governors.Controllers.AddOrEdit
                     eLookupGovernorRole.ChairOfGovernors,
                     false,
                     It.IsAny<IPrincipal>()))
-                .Returns(Task.FromResult(new GovernorDisplayPolicy()));
+                .ReturnsAsync(new GovernorDisplayPolicy());
 
             mockGovernorsWriteService
                 .Setup(g => g.ValidateAsync(It.IsAny<GovernorModel>(), It.IsAny<IPrincipal>()))
-                .Returns(Task.FromResult(new ValidationEnvelopeDto()));
+                .ReturnsAsync(new ValidationEnvelopeDto());
 
             mockGovernorsWriteService
                 .SetupSequence(g => g.SaveAsync(It.IsAny<GovernorModel>(), It.IsAny<IPrincipal>()))
-                .Returns(Task.FromResult(new ApiResponse<int>(true) { Response = 500 })) // first save
-                .Returns(Task.FromResult(new ApiResponse<int>(true) { Response = 501 })); // second save (reinstated governor)
+                .ReturnsAsync(new ApiResponse<int>(true) { Response = 500 })
+                .ReturnsAsync(new ApiResponse<int>(true) { Response = 501 });
 
             mockGovernorsWriteService
                 .Setup(g => g.UpdateDatesAsync(It.IsAny<int>(), It.IsAny<DateTime>(), It.IsAny<IPrincipal>()))
-                .Returns(Task.FromResult(new ApiResponse(true)));
+                .ReturnsAsync(new ApiResponse(true));
 
             mockGovernorsReadService
                 .Setup(g => g.GetGovernorAsync(oldChairId, It.IsAny<IPrincipal>()))
-                .Returns(Task.FromResult(new GovernorModel
+                .ReturnsAsync(new GovernorModel
                 {
                     Id = oldChairId,
                     AppointmentEndDate = oldChairEndDate
-                }));
+                });
 
             mockLayoutHelper
                 .Setup(l => l.PopulateLayoutProperties(
@@ -1784,6 +1944,8 @@ namespace Edubase.Web.UIUnitTests.Areas.Governors.Controllers.AddOrEdit
                     It.IsAny<Action<EstablishmentModel>>(),
                     It.IsAny<Action<GroupModel>>()))
                 .Returns(Task.CompletedTask);
+
+            SetupCommonLookupMocks();
 
             // Act
             var result = await controller.AddEditOrReplace(vm);
@@ -1804,55 +1966,117 @@ namespace Edubase.Web.UIUnitTests.Areas.Governors.Controllers.AddOrEdit
             Assert.IsType<RedirectResult>(result);
         }
 
-        [Theory()]
+        [Theory]
         [InlineData(null, 153513)]
         [InlineData(16851, null)]
         public async Task Gov_Post_AddEditOrReplace_ApiError(int? estabId, int? groupId)
         {
+            // Arrange
             var errorKey = "test";
             var errorMessage = "test message";
+            var controller = BuildController();
 
-            mockGovernorsReadService.Setup(g => g.GetEditorDisplayPolicyAsync(It.IsAny<eLookupGovernorRole>(), It.IsAny<bool>(), It.IsAny<IPrincipal>())).ReturnsAsync(() => new GovernorDisplayPolicy());
-            mockLayoutHelper.Setup(l => l.PopulateLayoutProperties(It.IsAny<CreateEditGovernorViewModel>(), It.IsAny<int?>(), It.IsAny<int?>(), It.IsAny<IPrincipal>(), It.IsAny<Action<EstablishmentModel>>(), It.IsAny<Action<GroupModel>>())).Returns(Task.CompletedTask);
+            SetupCommonLookupMocks();
+
+            mockGovernorsReadService
+                .Setup(g => g.GetEditorDisplayPolicyAsync(
+                    It.IsAny<eLookupGovernorRole>(),
+                    It.IsAny<bool>(),
+                    It.IsAny<IPrincipal>()))
+                .ReturnsAsync(() => new GovernorDisplayPolicy());
+
+            mockLayoutHelper
+                .Setup(l => l.PopulateLayoutProperties(
+                    It.IsAny<CreateEditGovernorViewModel>(),
+                    It.IsAny<int?>(),
+                    It.IsAny<int?>(),
+                    It.IsAny<IPrincipal>(),
+                    It.IsAny<Action<EstablishmentModel>>(),
+                    It.IsAny<Action<GroupModel>>()))
+                .Returns(Task.CompletedTask);
+
             mockGovernorsWriteService
                 .Setup(g => g.ValidateAsync(It.IsAny<GovernorModel>(), It.IsAny<IPrincipal>()))
-                .ReturnsAsync(() => new ValidationEnvelopeDto()
+                .ReturnsAsync(() => new ValidationEnvelopeDto
                 {
                     Errors =
                     {
-                        new ApiError {Code = "Test", Fields = errorKey, Message = errorMessage}
+                    new ApiError { Code = "Test", Fields = errorKey, Message = errorMessage }
                     }
                 });
 
-            var result = await controller.AddEditOrReplace(new CreateEditGovernorViewModel { EstablishmentUrn = estabId, GroupUId = groupId });
+            var vm = new CreateEditGovernorViewModel
+            {
+                EstablishmentUrn = estabId,
+                GroupUId = groupId
+            };
 
-            var viewResult = result as ViewResult;
-            Assert.NotNull(viewResult);
-            var model = viewResult.Model as CreateEditGovernorViewModel;
-            Assert.NotNull(model);
+            // Act
+            var result = await controller.AddEditOrReplace(vm);
+
+            // Assert
+            var viewResult = Assert.IsType<ViewResult>(result);
+            var model = Assert.IsType<CreateEditGovernorViewModel>(viewResult.Model);
+
             Assert.False(viewResult.ViewData.ModelState.IsValid);
             Assert.True(viewResult.ViewData.ModelState.ContainsKey(errorKey));
             Assert.Single(viewResult.ViewData.ModelState[errorKey].Errors);
             Assert.Equal(errorMessage, viewResult.ViewData.ModelState[errorKey].Errors[0].ErrorMessage);
         }
 
-        protected virtual void Dispose(bool disposing)
+        [Fact]
+        public async Task CreateGovernorsViewModel_BuildsBaseViewModel()
         {
-            if (!disposedValue)
-            {
-                if (disposing)
+            // Arrange
+            var estabUrn = 1000;
+
+            mockGovernorsReadService
+                .Setup(s => s.GetGovernorListAsync(estabUrn, null, It.IsAny<IPrincipal>()))
+                .ReturnsAsync(new GovernorsDetailsDto
                 {
-                    controller.Dispose();
-                }
+                    ApplicableRoles = new List<eLookupGovernorRole>(),
+                    CurrentGovernors = new List<GovernorModel>(),
+                    HistoricalGovernors = new List<GovernorModel>(),
+                    RoleDisplayPolicies = new Dictionary<eLookupGovernorRole, GovernorDisplayPolicy>()
+                });
 
-                disposedValue = true;
-            }
-        }
+            mockGovernorsReadService
+                .Setup(s => s.GetGovernorPermissions(estabUrn, null, It.IsAny<IPrincipal>()))
+                .ReturnsAsync(new GovernorPermissions());
 
-        public void Dispose()
-        {
-            Dispose(disposing: true);
-            GC.SuppressFinalize(this);
+            mockCachedLookupService.Setup(s => s.NationalitiesGetAllAsync())
+                .ReturnsAsync(new List<LookupDto>());
+
+            mockCachedLookupService.Setup(s => s.GovernorAppointingBodiesGetAllAsync())
+                .ReturnsAsync(new List<LookupDto>());
+
+            mockCachedLookupService.Setup(s => s.TitlesGetAllAsync())
+                .ReturnsAsync(new List<LookupDto>());
+
+            mockEstablishmentReadService
+                .Setup(s => s.GetAsync(1000, It.IsAny<IPrincipal>()))
+                .ReturnsAsync(new ServiceResultDto<EstablishmentModel>
+                {
+                    ReturnValue = new EstablishmentModel { GovernanceModeId = 1 },
+                    Status = eServiceResultStatus.Success
+                });
+
+            mockEstablishmentReadService
+                .Setup(s => s.GetPermissibleLocalGovernorsAsync(1000, It.IsAny<IPrincipal>()))
+                .ReturnsAsync(new List<LookupDto>());
+
+            // Act
+            var controller = BuildController();
+            var vm = await controller.CreateGovernorsViewModel(
+                groupUId: null,
+                establishmentUrn: estabUrn,
+                establishmentModel: null,
+                user: new GenericPrincipal(new GenericIdentity("UnitTestUser"), new string[0])
+                );
+
+            // Assert
+            Assert.NotNull(vm);
+            Assert.Equal(estabUrn, vm.EstablishmentUrn);
         }
     }
 }
