+++ Web/Edubase.Web.UIUnitTests/Areas/Governors/Models/GovernorViewModelTests.cs
@@ -307,5 +307,6 @@ namespace Edubase.Web.UIUnitTests.Areas.Governors.Models
         }
 
         #endregion
+
     }
 }
